[ARQ-1788] containers - code refactoring
- Create an ancestor for managed/remote containers using JMX
- Rename class OSGiContainerConfiguration to EmbeddedContainerConfiguration
- Process bootstrapCompleteServices in CommonDeployableContainer
