+++ container/common/src/main/java/org/jboss/arquillian/container/osgi/jmx/MBeanProxy.java
@@ -19,7 +19,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.jboss.arquillian.container.osgi.karaf.managed;
+package org.jboss.arquillian.container.osgi.jmx;
 
 import javax.management.MBeanServerConnection;
 import javax.management.MBeanServerInvocationHandler;
